# backend/swagger.yaml
swagger: "2.0"
info:
  description: "Trainer REST API documentation"
  version: "1.0.0"
  title: "Trainer"
host: "trainer.ksi.fit.cvut.cz"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /lessons/{id}:
    get:
      tags:
        - "lessons"
      summary: "Get detail of lesson with given id (including modules)"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, lesson detail"
          schema:
            $ref: "#/definitions/LessonGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to see lesson detail"
        "404":
          description: "Lesson with given identifier not found"
      security:
        - login_secret: [ ]
    patch:
      tags:
        - "lessons"
      summary: "Edit existing lesson"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson id"
          name: "id"
          type: "string"
          required: true
        - in: "body"
          name: "body"
          description: "Lesson values to be edited"
          required: true
          schema:
            $ref: "#/definitions/LessonUpdateDTO"
      responses:
        "200":
          description: "Success, lesson edited"
          schema:
            $ref: "#/definitions/LessonGetDTO"
        "403":
          description: "Insufficient rights to edit the lesson"
        "404":
          description: "Lesson with given identifier not found"
      security:
        - login_secret: []
    delete:
      tags:
        - "lessons"
      summary: "Delete existing lesson"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "string"
          required: true
      responses:
        "204":
          description: "Success, lesson deleted"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to delete the lesson"
        "404":
          description: "Lesson with given identifier not found"
      security:
        - login_secret: []
...
  /lessons/{id}/users:
    get:
      tags:
        - "lesson-users"
      summary: "Get lesson user list (with modules)"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, lesson user list (with modules)"
          schema:
            $ref: "#/definitions/LessonUserListDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to see lesson users"
        "404":
          description: "Lesson with given identifier not found"
      security:
        - login_secret: [ ]
...
  /lessons/{id}/modules/{moduleId}/users:
    get:
      tags:
        - "lesson-modules"
      summary: "Get LessonModule users list"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Unique module identifier"
          name: "moduleId"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, LessonModule users list"
          schema:
            $ref: "#/definitions/LessonModuleUserListDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to see LessonModule users"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]
  /lessons/{id}/modules:
    post:
      tags:
        - "lesson-modules"
      summary: "Copy module to lesson"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Description of copied module and new order"
          required: true
          schema:
            $ref: "#/definitions/LessonModuleCopyDTO"
      responses:
        "200":
          description: "Success, newly created LessonModule"
          schema:
            $ref: "#/definitions/LessonModuleReadDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to access lesson or module"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]
  /lessons/{id}/modules/{moduleId}:
    put:
      tags:
        - "lesson-modules"
      summary: "Add module to lesson"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Identifier of module to add / modify"
          name: "moduleId"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Description of LessonModule properties"
          required: true
          schema:
            $ref: "#/definitions/LessonModuleEditDTO"
      responses:
        "200":
          description: "Success, created/updated LessonModule"
          schema:
            $ref: "#/definitions/LessonModuleReadDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to access lesson or module"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]
    delete:
      tags:
        - "lesson-modules"
      summary: "Remove module from lesson"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Identifier of module to remove"
          name: "moduleId"
          type: "integer"
          required: true
      responses:
        "204":
          description: "Success, removed LessonModule"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to access lesson or module"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]

  # Modules
...
  /modules/{id}:
    get:
      tags:
        - "modules"
      summary: "Get module detail"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique module identifier"
          name: "id"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, module detail"
          schema:
            $ref: "#/definitions/ModuleGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to view the module"
        "404":
          description: "Module with given identifier not found"
      security:
        - login_secret: []
    patch:
      tags:
        - "modules"
      summary: "Edit existing module"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique module identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Module values to be updated"
          required: true
          schema:
            $ref: "#/definitions/ModuleUpdateDTO"
      responses:
        "200":
          description: "Success, module updated"
          schema:
            $ref: "#/definitions/ModuleGetDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to edit the module"
        "404":
          description: "Module with given identifier not found"
      security:
        - login_secret: []
    delete:
      tags:
        - "modules"
      summary: "Delete existing module"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique module identifier"
          name: "id"
          type: "integer"
          required: true
      responses:
        "204":
          description: "Success, module deleted"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to delete the module"
        "404":
          description: "Module with given identifier not found"
      security:
        - login_secret: []
...
securityDefinitions:
  login_secret:
    type: "apiKey"
    name: "LoginSecret"
    in: "header"
definitions:
...
  CourseFindDTO:
    description: "Course list DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique course identifier"
        example: 1
      name:
        type: "string"
        description: "Course name"
        example: "Programming and Algorithms 1 (3 + 4)"
      subject:
        $ref: "#/definitions/SubjectFindDTO"
      semester:
        $ref: "#/definitions/SemesterFindDTO"
      lessonCompleted:
        type: "integer"
        description: "Number of completed lessons in course"
        example: 2
      lessonCount:
        type: "integer"
        description: "Number of lessons in course"
        example: 5
      role:
        $ref: "#/definitions/RoleEnum"
...
  SubjectFindDTO:
    description: "Subject DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique subject identifier"
        example: 1
      name:
        type: "string"
        description: "Subject name"
        example: "Programming and Algorithms 1"
      code:
        type: "string"
        description: "Subject code"
        example: "BI-PA1.21"
  SemesterFindDTO:
    description: "Semester DTO"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique semester identifier"
        example: 1
      code:
        type: "string"
        description: "Semester code"
        example: "B232"
      from:
        type: "string"
        description: "Semester start date"
        format: "date-time"
      until:
        type: "string"
        description: "Semester end date"
        format: "date-time"

  # Week
  WeekFindDTO:
    description: "Week"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique week identifier"
        example: 1
      name:
        type: "string"
        description: "Week name"
        example: "1st week"
      from:
        type: "string"
        format: "date-time"
        description: "Week start date"
        example: "2024-02-01T00:00:00.000+00:00"
      until:
        type: "string"
        format: "date-time"
        description: "Week end date"
        example: "2025-02-01T00:00:00.000+00:00"
      course:
        $ref: "#/definitions/CourseFindDTO"
...
  CourseUserReadDTO:
    description: "Course user"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique user identifier"
        example: 1
      username:
        type: "string"
        description: "User's username"
        example: "wrzecond"
      name:
        type: "string"
        description: "User's name"
        example: "Ondřej Wrzecionko"
      role:
        $ref: "#/definitions/RoleEnum"
      progress:
        type: "integer"
        description: "User progress in course (in percent)"
        example: 24
...
  LessonDTO:
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique lesson identifier"
        example: 1
      name:
        type: "string"
        description: "Lesson name"
        example: "Compilation"
      hidden:
        type: "boolean"
        description: "Is lesson hidden from students?"
        example: false
      type:
        $ref: "#/definitions/LessonType"
      lockCode:
        type: "string"
        description: "Code for unlocking modules in lesson (teacher only)"
        example: null
      timeLimit:
        type: "string"
        format: "date-time"
        description: "Lesson end date"
        example: "2025-02-01T00:00:00.000+00:00"
...
  LessonGetDTO:
    description: "Lesson"
    type: "object"
    allOf:
      - $ref: "#/definitions/LessonDTO"
    properties:
      week:
        $ref: "#/definitions/WeekFindDTO"
      timeStart:
        type: "string"
        format: "date-time"
        description: "Lesson start date"
        example: "2024-02-01T00:00:00.000+00:00"
      description:
        type: "string"
        description: "Lesson description (introduction text)"
        example: "In this lesson, we will cover the basics of how to compile..."
      modules:
        $ref: "#/definitions/ModuleFindDTOArray"
  LessonType:
    type: "string"
    enum: [ "TUTORIAL", "TUTORIAL_PREPARATION", "INDIVIDUAL_TASK", "SUPPLEMENTARY", "FREQUENT_MISTAKES", "INFORMATION" ]
    description: "Lesson type"
    example: "TUTORIAL"
  LessonUpdateDTO:
    type: "object"
    description: "DTO used for updating lesson"
    properties:
      name:
        type: "string"
        description: "Newly created lesson name"
        example: "Another lesson"
      hidden:
        type: "boolean"
        description: "Should lesson be hidden from students?"
        example: false
      order:
        type: "integer"
        description: "Lesson order in week"
        example: 2
      type:
        $ref: "#/definitions/LessonType"
      lockCode:
        type: "string"
        description: "Code used for unlocking lesson modules"
        example: "secretcode"
      timeStart:
        type: "string"
        format: "date-time"
        description: "Lesson start date"
        example: "2024-02-01T00:00:00.000+00:00"
      timeLimit:
        type: "string"
        format: "date-time"
        description: "Time limit for user to solve lesson in time"
        example: "2025-02-01T00:00:00.000+00:00"
      description:
        type: "string"
        description: "Lesson description (with assignment, conditions...)"
        example: "In this lesson, we will cover the basics of how to compile..."
...
  LessonUserListDTO:
    type: "object"
    description: "DTO used for lesson user list"
    properties:
      id:
        type: "integer"
        description: "Lesson id"
        example: 1
      course:
        $ref: "#/definitions/CourseFindDTO"
      name:
        type: "string"
        description: "Lesson name"
        example: "Compilation"
      modules:
        type: "array"
        description: "Modules list"
        items:
          $ref: "#/definitions/ModuleFindDTO"
      users:
        type: "array"
        description: "Users with progress list"
        items:
          $ref: "#/definitions/ModuleUserListDTO"
...
  LessonModuleUserListDTO:
    type: "object"
    description: "LessonModule users list (1 lesson, 1 module, all solutions)"
    properties:
      lesson:
        $ref: "#/definitions/LessonGetDTO"
      module:
        $ref: "#/definitions/ModuleFindDTO"
      users:
        type: "array"
        description: "Users with progress in module list"
        items:
          $ref: "#/definitions/ModuleUserSingleListDTO"
  ModuleUserListDTO:
    type: "object"
    description: "User with list of module progress"
    properties:
      id:
        type: "integer"
        description: "User id"
        example: 1
      username:
        type: "string"
        description: "Username"
        example: "wrzecond"
      name:
        type: "string"
        description: "User's name"
        example: "Ondřej Wrzecionko"
      modules:
        type: "array"
        description: "Array of user – module progress"
        items:
          $ref: "#/definitions/ModuleUserDTO"
  ModuleUserSingleListDTO:
    type: "object"
    description: "User with module progress"
    properties:
      id:
        type: "integer"
        description: "User id"
        example: 1
      username:
        type: "string"
        description: "Username"
        example: "wrzecond"
      name:
        type: "string"
        description: "User's name"
        example: "Ondřej Wrzecionko"
      module:
        $ref: "#/definitions/ModuleUserDTO"
  ModuleUserDTO:
    type: "object"
    description: "User's progress in one module"
    properties:
      id:
        type: "integer"
        description: "Module identifier"
        example: 42
      requestType:
        $ref: "#/definitions/StudentModuleRequestType"
      allowedShow:
        type: "boolean"
        description: "Has student allowed anonymized publication?"
        example: true
      progress:
        type: "integer"
        description: "User progress in module (in percent, 0 – 100)"
        example: 22
      completedEarly:
        type: "boolean"
        description: "Has student completed module before lesson end?"
        example: false
      completed:
        type: "string"
        format: "date-time"
        description: "Time when user has completed the module (null if uncompleted)"
        example: "2024-02-19T12:11:05.000+00:00"

  # Lesson module
  LessonModuleReadDTO:
    type: "object"
    description: "LessonModule description"
    properties:
      id:
        type: "integer"
        description: "Unique identifier of LessonModule"
        example: 44
      name:
        type: "string"
        description: "Name of module"
        example: "Compilation – task"
      type:
        type: "string"
        enum: [ "TEXT", "QUIZ", "CODE", "ASSIGNMENT" ]
        description: "Type of module"
        example: "ASSIGNMENT"
      order:
        type: "integer"
        description: "Module order in lesson"
        example: 1
  LessonModuleEditDTO:
    type: "object"
    description: "Description of module to edit / create"
    properties:
      order:
        type: "integer"
        description: "Module order in lesson"
        example: 1
      depends:
        type: "integer"
        description: "Identifier of module on which the module should depend (null if none)"
        example: 2
  LessonModuleCopyDTO:
    type: "object"
    description: "Description of module to copy"
    properties:
      copiedId:
        type: "integer"
        description: "Unique identifier of module to copy"
        example: 12
      order:
        type: "integer"
        description: "Module order in new lesson"
        example: 1

  # Module
  ModuleFindDTOArray:
    type: "array"
    items:
      $ref: "#/definitions/ModuleFindDTO"
  ModuleFindDTO:
    allOf:
      - $ref: "#/definitions/ModuleGetDTO"
    properties:
      locked:
        type: "boolean"
        description: "Flag determining, whether module is really locked"
        example: false
      file:
        type: "string"
        description: "If module has file (test data), it will be the name of the file"
        example: "d32kvcz21y"
      author:
        type: "string"
        description: "Module author"
        example: "wrzecond"
      editors:
        type: "array"
        description: "Module editors (usernames)"
        items:
          type: "string"
        example: ["matouj10", "pasekmat"]
      allowedShow:
        type: "boolean"
        description: "Flag determining whether student allowed anonymized publishing of the module"
        example: false
      completedEarly:
        type: "boolean"
        description: "Flag determining whether student completed the module before lesson end"
        example: true
      completed:
        type: "string"
        format: "date-time"
        description: "Datetime when student completed the module (null if not completed)"
        example: "2024-03-01T14:16:11.000+00:00"
      requestType:
        type: "string"
        enum: ["HELP", "EVALUATE"]
        description: "Type of request that is currently active"
        example: "EVALUATE"
      teacherComment:
        type: "string"
        description: "Teacher's comment on the last student request"
        example: "It looks good, but try better."
      requestText:
        type: "string"
        description: "Student's comment associated with current active request"
        example: "I improved it, please evaluate it."
      order:
        type: "integer"
        description: "Module order in lesson"
        example: 2
      progress:
        type: "integer"
        description: "Student progress in module (in percent)"
        example: 100
      hasStudentFile:
        type: "boolean"
        description: "Flag indicating whether module has a solution (file) uploaded by student"
        example: true
  ModuleGetDTO:
    description: "Module"
    type: "object"
    properties:
      id:
        type: "integer"
        description: "Unique module identifier"
        example: 1
      name:
        type: "string"
        description: "Module name"
        example: "Assignment: Compile on your own computer"
      depends:
        type: "integer"
        description: "ID of module this module is depending on"
        example: 2
      author:
        $ref: "#/definitions/CourseUserReadDTO"
      editors:
        type: "array"
        description: "IDs of users who are editors of the module"
        items:
          type: "integer"
      type:
        type: "string"
        enum: ["TEXT", "QUIZ", "CODE", "ASSIGNMENT"]
        description: "Type of module"
        example: "CODE"
      lastModificationTime:
        type: "string"
        format: "date-time"
        description: "Last time, when module was changed"
        example: "2024-02-13T11:03:20.000+00:00"
      difficulty:
        type: "string"
        enum: ["BEGINNER", "EASY", "MEDIUM", "DIFFICULT", "EXTREME"]
        description: "Module difficulty rating"
        example: "EASY"
      assignment:
        type: "string"
        description: "Markdown assignment of module"
        example: "In this task, you have to compile the following code on your computer and upload it."
      minPercent:
        type: "integer"
        description: "Minimum percent needed for completing the module"
        example: 100
      lockable:
        type: "boolean"
        description: "Flag determining, whether module can be locked."
        example: false
      timeLimit:
        type: "boolean"
        description: "Flag determining, whether module should be solved 'early'."
        example: true
      manualEval:
        type: "boolean"
        description: "Flag determining, whether module should be evaluated manually (by teacher)."
        example: true
      hidden:
        type: "boolean"
        description: "Flag determining, whether module is hidden from students."
        example: false
  ModuleEditDTO:
    description: "Module"
    type: "object"
    properties:
      name:
        type: "string"
        description: "Module name"
        example: "Assignment: Compile on your own computer"
      editors:
        type: "array"
        description: "IDs of users who are editors of the module"
        items:
          type: "integer"
          example: 1
      difficulty:
        type: "string"
        enum: [ "BEGINNER", "EASY", "MEDIUM", "DIFFICULT", "EXTREME" ]
        description: "Module difficulty rating"
        example: "EASY"
      assignment:
        type: "string"
        description: "Markdown assignment of module"
        example: "In this task, you have to compile the following code on your computer and upload it."
      minPercent:
        type: "integer"
        description: "Minimum percent needed for completing the module"
        example: 100
      lockable:
        type: "boolean"
        description: "Flag determining, whether module can be locked."
        example: false
      timeLimit:
        type: "boolean"
        description: "Flag determining, whether module should be solved 'early'."
        example: true
      manualEval:
        type: "boolean"
        description: "Flag determining, whether module should be evaluated manually (by teacher)."
        example: true
      hidden:
        type: "boolean"
        description: "Flag determining, whether module is hidden from students."
        example: false
...
  ModuleUpdateDTO:
    description: "Module update DTO"
    allOf:
      - $ref: "#/definitions/ModuleEditDTO"
    properties:
      lastModificationTime:
        type: "string"
        format: "date-time"
        description: "Last time, when module was changed"
        example: "2024-02-13T11:03:20.000+00:00"

  # Student module
  StudentModuleRequestType:
    type: "string"
    description: "Student request type"
    enum: ["HELP", "COMMENT", "EVALUATE"]
    example: "HELP"
...
  RoleEnum:
    type: "string"
    description: "User role in course"
    enum: ["TEACHER", "STUDENT"]
    example: "STUDENT"

  # User
...